+++ repos/core/src/pages/eis/components/TableColGroup.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const TableColGroup = ({ colGroup }) => {
+  return (
+    <colgroup>
+      {colGroup.map(column => {
+        return <col style={column.style}/>
+      })}
+    </colgroup>
+  )
+}
+
+export default TableColGroup;